Give possibility to make glibc PRNG work for every seed
Because of the modulo trick the current glibc PRNG version doesn't work for seed == 0x7fffffff or seed == 0xfffffffe.

The former is used only if the user specifically uses --start / --end 02/2038. The latter is never used as negative values are not legal for time.

In any case a compilation guard (PWPS_UNERRING) has been added for correctness to make the PRNG spit correct values even for those 2 seeds.